Mislabeled request
It appears that one of the requests was copy-pasted from somewhere else; the request is a good one, but the description was entirely wrong. 

I put a generic description since all I know is what the API returns when I use this request; perhaps someone with inside knowledge could better qualify what this endpoint does.